webcomp/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
html5lib==1.1
idna==2.10
imageio==2.9.0
imgcompare==2.0.1
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.17.2
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.2
tifffile==2020.9.3
urllib3==1.26.3
webencodings==0.5.1
dnspython==2.1.0
py3dns==3.2.1